+++ pkg/cmds/parameters/parsed-parameter.go
@@ -1,10 +1,7 @@
 package parameters
 
 import (
-	reflect2 "github.com/go-go-golems/glazed/pkg/helpers/reflect"
-	"github.com/pkg/errors"
 	"github.com/wk8/go-ordered-map/v2"
-	"reflect"
 )
 
 type ParsedParameter struct {
@@ -201,75 +198,4 @@ func (p *ParsedParameters) ToMap() map[string]interface{} {
 		ret[k] = v.Value
 	})
 	return ret
-}
-
-// InitializeStruct initializes a struct from a ParsedParameters map.
-//
-// It iterates through the struct fields looking for those tagged with
-// "glazed.parameter". For each tagged field, it will lookup the corresponding
-// parameter value in the ParsedParameters map and set the field's value.
-//
-// Struct fields that are pointers to other structs are handled recursively.
-//
-// s should be a pointer to the struct to initialize.
-//
-// ps is the ParsedParameters map to lookup parameter values from.
-//
-// Returns an error if:
-// - s is not a pointer to a struct
-// - A tagged field does not have a matching parameter value in ps
-// - Failed to set the value of a field
-func (p *ParsedParameters) InitializeStruct(s interface{}) error {
-	if s == nil {
-		return errors.Errorf("Can't initialize nil struct")
-	}
-	// check that s is indeed a pointer to a struct
-	if reflect.TypeOf(s).Kind() != reflect.Ptr {
-		return errors.Errorf("s is not a pointer")
-	}
-	if reflect.TypeOf(s).Elem().Kind() != reflect.Struct {
-		return errors.Errorf("s is not a pointer to a struct")
-	}
-	st := reflect.TypeOf(s).Elem()
-
-	for i := 0; i < st.NumField(); i++ {
-		field := st.Field(i)
-		v, ok := field.Tag.Lookup("glazed.parameter")
-		if !ok {
-			continue
-		}
-		v_, ok := p.Get(v)
-		if !ok {
-			continue
-		}
-		value := reflect.ValueOf(s).Elem().FieldByName(field.Name)
-
-		if field.Type.Kind() == reflect.Ptr {
-			elem := field.Type.Elem()
-			if value.IsNil() {
-				value.Set(reflect.New(elem))
-			}
-			//exhaustive:ignore
-			switch elem.Kind() {
-			case reflect.Struct:
-				err := p.InitializeStruct(value.Interface())
-				if err != nil {
-					return errors.Wrapf(err, "failed to initialize struct for %s", v)
-				}
-			default:
-				err := reflect2.SetReflectValue(value.Elem(), v_.Value)
-				if err != nil {
-					return errors.Wrapf(err, "failed to set value for %s", v)
-				}
-			}
-
-		} else {
-			err := reflect2.SetReflectValue(value, v_.Value)
-			if err != nil {
-				return errors.Wrapf(err, "failed to set value for %s", v)
-			}
-		}
-	}
-
-	return nil
 }
